+++ src/app/auth/bets/create/page.tsx
@@ -3,7 +3,7 @@ import { CreateBetForm } from '@/features/bets/components/CreateBetForm';
 
 export default async function CreateBet() {
   return (
-    <main className='flex flex-col xs:px-4 lg:px-default flex-1 overflow-y-scroll items-center xs:justify-start lg:justify-center'>
+    <main className='flex flex-col xs:px-4 lg:px-default py-4 flex-1 overflow-y-scroll items-center xs:justify-start lg:justify-center'>
       <CreateBetForm />
     </main>
   );