+++ src/components/AppHeader.tsx
@@ -9,6 +9,7 @@ import { useEffect, useMemo, useRef, useState } from 'react';
 import { IconButton } from '@mui/material';
 import pkg from 'betting_app/package.json';
 import { Logo } from './Logo';
+import { ToggleProvider } from './feature/ToggleProvider';
 
 export function AppHeader() {
   const { user, status } = useUserContext();
@@ -35,35 +36,41 @@ export function AppHeader() {
 
   const MenuIcon = menuOpen ? Clear : Menu;
   return (
-    <div className='flex flex-col w-full relative'>
-      <Header ref={headerRef}>
-        <div className='flex justify-between items-center w-full'>
-          <div className='flex flex-col'>
-            <Link
-              className='flex gap-2 items-center index-link'
-              href='/'>
-              <Logo />
-            </Link>
-            <small className='ml-8'>{pkg.version}</small>
-          </div>
+    <ToggleProvider onChange={state => setMenuOpen(state)}>
+      <div className='flex flex-col w-full relative'>
+        <Header ref={headerRef}>
+          <div className='flex justify-between items-center w-full'>
+            <div className='flex flex-col'>
+              <Link
+                className='flex gap-2 items-center index-link'
+                href='/'>
+                <Logo />
+              </Link>
+              <small className='ml-8'>{pkg.version}</small>
+            </div>
 
-          <nav className='gap-8 xs:hidden lg:flex'>{getLinks()}</nav>
-          <div className='lg:hidden xs:block'>
-            <IconButton onClick={() => setMenuOpen(prev => !prev)}>
-              <MenuIcon sx={{ color: 'var(--color-accent)', zIndex: 30, margin: 0, padding: 0 }} />
-            </IconButton>
+            <nav className='gap-8 xs:hidden lg:flex'>{getLinks()}</nav>
+            <div className='lg:hidden xs:block'>
+              <ToggleProvider.Trigger>
+                <IconButton>
+                  <MenuIcon
+                    sx={{ color: 'var(--color-accent)', zIndex: 30, margin: 0, padding: 0 }}
+                  />
+                </IconButton>
+              </ToggleProvider.Trigger>
+            </div>
           </div>
-        </div>
-      </Header>
-      {menuOpen && (
-        <nav
-          onClick={() => setMenuOpen(false)}
-          id='drop-down-menu'
-          className='absolute w-full flex flex-col items-end gap-8 z-20 bg-white p-4 border-b border-border animate-slide-down'
-          style={{ top: headerRef.current?.offsetHeight || 0 }}>
-          {getLinks()}
-        </nav>
-      )}
-    </div>
+        </Header>
+
+        <ToggleProvider.Target hideOnClickOutside>
+          <nav
+            id='drop-down-menu'
+            className='absolute w-full flex flex-col items-end gap-8 z-20 bg-white p-4 border-b border-border animate-slide-down'
+            style={{ top: headerRef.current?.offsetHeight || 0 }}>
+            {getLinks()}
+          </nav>
+        </ToggleProvider.Target>
+      </div>
+    </ToggleProvider>
   );
 }